Hide close button on full page
Move header banner to PortalBanner

Get chainId dand controller via context

Introduce layout component submodule

Add content layout component

Add layout context

Style connect variant banner
