+++ examples/starknet-react-next/src/components/StarknetProvider.tsx
@@ -52,7 +52,7 @@ const connectors = [
     ],
     {
       url,
-      rpc: process.env.NEXT_PUBLIC_RPC_SEPOLIA
+      rpc: process.env.NEXT_PUBLIC_RPC_SEPOLIA,
       // theme: "rollyourown",
       // colorMode: "light"
     },
