+++ packages/keychain/src/components/DeploymentRequired.tsx
@@ -1,26 +1,22 @@
 import { constants } from "starknet";
-import Controller from "utils/controller";
-import { Container } from "./Container";
+import { Container, Banner, Footer, Content } from "components/layout";
 import { useEffect, useState } from "react";
 import { Status } from "utils/account";
 import { Loading } from "./Loading";
 import { Button, Link, Text } from "@chakra-ui/react";
 import { ExternalIcon } from "@cartridge/ui";
-import { PortalBanner } from "./PortalBanner";
-import { PortalFooter } from "./PortalFooter";
+import { useController } from "hooks/controller";
+
 export function DeploymentRequired({
-  chainId,
-  controller,
   onClose,
   onLogout,
   children,
 }: {
-  chainId: string;
-  controller: Controller;
   onClose: () => void;
   onLogout: () => void;
   children: React.ReactNode;
 }) {
+  const { controller } = useController()
   const account = controller.account;
   const [status, setStatus] = useState<Status>(account.status);
   const [deployHash, setDeployHash] = useState<string>();
@@ -65,40 +61,41 @@ export function DeploymentRequired({
 
   if (status !== Status.DEPLOYED) {
     return (
-      <Container chainId={chainId} onLogout={onLogout}>
-        <PortalBanner
+      <Container onLogout={onLogout}>
+        <Banner
           Icon={Loading}
           title={"Deploying your account"}
           description="This may take a second"
         />
 
-        {status === Status.DEPLOYING && (
-          <Link
-            href={`https://${
-              account.chainId === constants.StarknetChainId.SN_SEPOLIA
+        <Content>
+          {status === Status.DEPLOYING && (
+            <Link
+              href={`https://${account.chainId === constants.StarknetChainId.SN_SEPOLIA
                 ? "sepolia."
                 : undefined
-            }starkscan.co/tx/${deployHash}`}
-            isExternal
-          >
-            <Button variant="link" mt={10} rightIcon={<ExternalIcon />}>
-              View on Starkscan
-            </Button>
-          </Link>
-        )}
+                }starkscan.co/tx/${deployHash}`}
+              isExternal
+            >
+              <Button variant="link" mt={10} rightIcon={<ExternalIcon />}>
+                View on Starkscan
+              </Button>
+            </Link>
+          )}
 
-        {error && (
-          <>
-            <Text>
-              We encounter an account deployment error: {error.message}
-            </Text>
-            <Text>Please come by discord and report this issue.</Text>
-          </>
-        )}
+          {error && (
+            <>
+              <Text>
+                We encounter an account deployment error: {error.message}
+              </Text>
+              <Text>Please come by discord and report this issue.</Text>
+            </>
+          )}
+        </Content>
 
-        <PortalFooter>
+        <Footer>
           <Button onClick={onClose}>close</Button>
-        </PortalFooter>
+        </Footer>
       </Container>
     );
   }
