+++ packages/keychain/src/components/Execute/Fees.tsx
@@ -11,6 +11,7 @@ import {
 import { constants } from "starknet";
 import { formatUnits } from "viem";
 import { Error } from "components/Error";
+import { useChainId } from "hooks/connection";
 
 async function fetchEthPrice() {
   const res = await fetch(process.env.NEXT_PUBLIC_API_URL, {
@@ -25,17 +26,16 @@ async function fetchEthPrice() {
 
 export function Fees({
   error,
-  chainId,
   fees,
   balance,
   approved,
 }: {
   error: Error;
-  chainId: string;
   fees?: { base: bigint; max: bigint };
   balance: string;
   approved?: string;
 }) {
+  const chainId = useChainId();
   const [formattedFee, setFormattedFee] = useState<{
     base: string;
     max: string;
@@ -66,13 +66,13 @@ export function Fees({
       setFormattedFee(
         fees.max > 10000000000000n
           ? {
-              base: `~${parseFloat(formatUnits(fees.base, 18)).toFixed(5)} eth`,
-              max: `~${parseFloat(formatUnits(fees.max, 18)).toFixed(5)} eth`,
-            }
+            base: `~${parseFloat(formatUnits(fees.base, 18)).toFixed(5)} eth`,
+            max: `~${parseFloat(formatUnits(fees.max, 18)).toFixed(5)} eth`,
+          }
           : {
-              base: "<0.00001",
-              max: "<0.00001",
-            },
+            base: "<0.00001",
+            max: "<0.00001",
+          },
       );
     }
     compute();
