+++ packages/keychain/src/components/Execute/index.tsx
@@ -1,54 +1,51 @@
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { Text, VStack, Button } from "@chakra-ui/react";
-
-import Controller from "utils/controller";
 import { Call as StarknetCall, InvocationsDetails } from "starknet";
 import { Fees } from "./Fees";
 import { formatEther } from "viem";
 import { ExecuteReply, ResponseCodes } from "@cartridge/controller";
-import { Container } from "../Container";
+import {
+  Container,
+  Content,
+  Banner,
+  FOOTER_MIN_HEIGHT,
+  Footer,
+} from "components/layout";
 import { Status } from "utils/account";
-import { PortalBanner } from "../PortalBanner";
 import { TransactionDuoIcon } from "@cartridge/ui";
 import { Call } from "./Call";
-import {
-  PORTAL_FOOTER_MIN_HEIGHT,
-  PortalFooter,
-} from "components/PortalFooter";
-import LowEth from "./LowEth";
+import { InsufficientFunds } from "./InsufficientFunds";
+import { useChainId, useOrigin } from "hooks/connection";
+import { useController } from "hooks/controller";
 
 export const CONTRACT_ETH =
   "0x049d36570d4e46f48e99674bd3fcc84644ddd6b96f7c741b1562b82f9e004dc7";
 
 export function Execute({
-  // origin,
-  chainId,
-  controller,
   transactions,
   transactionsDetail,
-  // abis,
   onExecute,
   onCancel,
   onLogout,
 }: {
-  // origin: string;
-  chainId: string;
-  controller: Controller;
   transactions: StarknetCall | StarknetCall[];
   transactionsDetail?: InvocationsDetails;
-  // abis?: Abi[];
   onExecute: (res: ExecuteReply) => void;
   onCancel: () => void;
   onLogout: () => void;
 }) {
+  const chainId = useChainId();
+  const { controller } = useController();
+  const origin = useOrigin();
+
   const [fees, setFees] = useState<{
     base: bigint;
     max: bigint;
   }>();
   const [error, setError] = useState<Error>();
   const [isLoading, setLoading] = useState<boolean>(false);
-  const [ethBalance, setEthBalance] = useState<bigint>();
-  const [lowEth, setLowEth] = useState<boolean>(false);
+  const [ethBalance, setEthBalance] = useState<bigint>(0n);
+  const [isInsufficient, setIsInsufficient] = useState<boolean>(false);
 
   const account = controller.account;
   const calls = useMemo(() => {
@@ -122,7 +119,7 @@ export function Execute({
     }
 
     if (ethBalance < fees.max) {
-      setLowEth(true);
+      setIsInsufficient(true);
     }
   }, [ethBalance, fees]);
 
@@ -137,10 +134,9 @@ export function Execute({
     });
   }, [account, calls, fees, onExecute]);
 
-  if (lowEth) {
+  if (isInsufficient) {
     return (
-      <LowEth
-        chainId={chainId}
+      <InsufficientFunds
         address={controller.account.address}
         balance={format(ethBalance)}
       />
@@ -148,10 +144,10 @@ export function Execute({
   }
 
   return (
-    <Container chainId={chainId} onLogout={onLogout}>
-      <PortalBanner Icon={TransactionDuoIcon} title="Submit Transaction" />
+    <Container onLogout={onLogout}>
+      <Banner Icon={TransactionDuoIcon} title="Confirm Transaction" description={origin} />
 
-      <VStack w="full" pb={lowEth ? undefined : PORTAL_FOOTER_MIN_HEIGHT}>
+      <Content pb={FOOTER_MIN_HEIGHT}>
         <VStack spacing="1px" w="full" borderRadius="md" bg="solid.primary">
           <VStack w="full" p={3} align="flex-start">
             <Text fontSize="xs" fontWeight="bold" color="text.secondaryAccent">
@@ -163,7 +159,6 @@ export function Execute({
             {calls.map((call, i) => (
               <Call
                 key={i}
-                chainId={chainId}
                 policy={{
                   target: call.contractAddress,
                   method: call.entrypoint,
@@ -174,28 +169,25 @@ export function Execute({
           </VStack>
         </VStack>
 
-        <VStack w="full">
-          <Fees
-            error={error}
-            chainId={chainId}
-            fees={fees}
-            balance={ethBalance && format(ethBalance)}
-          />
+        <Fees
+          error={error}
+          fees={fees}
+          balance={ethBalance && format(ethBalance)}
+        />
+      </Content>
 
-          <PortalFooter>
-            <Button
-              colorScheme="colorful"
-              onClick={onSubmit}
-              isLoading={isLoading}
-              isDisabled={!fees}
-            >
-              submit
-            </Button>
+      <Footer>
+        <Button
+          colorScheme="colorful"
+          onClick={onSubmit}
+          isLoading={isLoading}
+          isDisabled={!fees}
+        >
+          submit
+        </Button>
 
-            <Button onClick={onCancel}>Cancel</Button>
-          </PortalFooter>
-        </VStack>
-      </VStack>
+        <Button onClick={onCancel}>Cancel</Button>
+      </Footer>
     </Container>
   );
 }
