--- packages/keychain/src/components/Quests/types.ts
@@ -1,5 +1,0 @@
-export enum QuestState {
-  Incomplete,
-  Claimable,
-  Complete,
-}
