+++ packages/keychain/src/components/connect/Login.tsx
@@ -1,12 +1,13 @@
 import { Field } from "@cartridge/ui";
-import { VStack, Button } from "@chakra-ui/react";
-import { Container } from "../Container";
-import { Form as FormikForm, Field as FormikField, Formik } from "formik";
+import { Button } from "@chakra-ui/react";
 import {
-  PORTAL_FOOTER_MIN_HEIGHT,
-  PortalBanner,
-  PortalFooter,
-} from "components";
+  Container,
+  FOOTER_MIN_HEIGHT,
+  Banner,
+  Footer,
+  Content,
+} from "components/layout";
+import { Form as FormikForm, Field as FormikField, Formik } from "formik";
 import { useCallback, useState } from "react";
 import Controller from "utils/controller";
 import { FormValues, LoginMode, LoginProps } from "./types";
@@ -104,7 +105,7 @@ export function Login({
   );
 
   return (
-    <Container chainId={chainId} overflowY={error ? "auto" : undefined}>
+    <Container variant="connect" overflowY={error ? "auto" : undefined}>
       <Formik
         initialValues={{ username: prefilledName }}
         onSubmit={onSubmit}
@@ -113,7 +114,7 @@ export function Login({
       >
         {(props) => (
           <FormikForm style={{ width: "100%" }}>
-            <PortalBanner
+            <Banner
               title={
                 theme.id === "cartridge"
                   ? "Play with Cartridge Controller"
@@ -122,10 +123,7 @@ export function Login({
               description="Enter your Controller username"
             />
 
-            <VStack
-              align="stretch"
-              pb={error ? PORTAL_FOOTER_MIN_HEIGHT : undefined}
-            >
+            <Content pb={error ? FOOTER_MIN_HEIGHT : undefined}>
               <FormikField
                 name="username"
                 placeholder="Username"
@@ -146,9 +144,9 @@ export function Login({
               </FormikField>
 
               <ErrorComp error={error} />
-            </VStack>
+            </Content>
 
-            <PortalFooter origin={origin} policies={policies} isSlot={isSlot}>
+            <Footer origin={origin} policies={policies} isSlot={isSlot} showLogo>
               <Button
                 type="submit"
                 colorScheme="colorful"
@@ -162,7 +160,7 @@ export function Login({
               >
                 Sign up
               </RegistrationLink>
-            </PortalFooter>
+            </Footer>
           </FormikForm>
         )}
       </Formik>
