+++ packages/keychain/src/components/connect/index.tsx
@@ -2,5 +2,5 @@ export * from "./Authenticate";
 export * from "./CreateController";
 export * from "./CreateSession";
 export * from "./Login";
+export * from "./Logout";
 export * from "./Signup";
-export * from "./StarterPack";
