+++ packages/keychain/src/components/index.ts
@@ -1,12 +1,6 @@
 export * from "./Animation";
-export * from "./connect";
-export * from "./Container";
 export * from "./DeploymentRequired";
 export * from "./Execute";
 export * from "./Loading";
-export * from "./Logout";
-export * from "./PortalBanner";
-export * from "./PortalFooter";
-export * from "./Quests";
 export * from "./Redeploy";
 export * from "./SignMessage";
