+++ packages/keychain/src/components/layout/Container/index.tsx
@@ -1,57 +1,32 @@
+export { TOP_OFFSET } from "./Header"
+
 import {
   Container as ChakraContainer,
   VStack,
   StyleProps,
   Flex,
   Show,
-  HStack,
-  Text,
-  IconButton,
 } from "@chakra-ui/react";
 import { motion } from "framer-motion";
-import { ReactNode } from "react";
+import { ReactNode, createContext, useContext } from "react";
 import { Header, HeaderProps } from "./Header";
-import { constants } from "starknet";
-import { CartridgeLogo, TimesIcon } from "@cartridge/ui";
-import { useConnection } from "hooks/connection";
-import { isIframe } from "components/connect/utils";
 
 export function Container({
   children,
-  chainId = constants.StarknetChainId.SN_SEPOLIA,
   onBack,
   hideAccount,
   ...rest
 }: {
+  variant?: LayoutVariant;
   children: ReactNode;
 } & StyleProps &
   HeaderProps) {
-  const { close } = useConnection();
-
   return (
     <Wrapper {...rest}>
-      {isIframe() && (
-        <IconButton
-          aria-label="Close Keychain"
-          icon={<TimesIcon />}
-          position="absolute"
-          zIndex="9999999"
-          colorScheme="translucent"
-          size="sm"
-          h={8}
-          top={3}
-          left={3}
-          onClick={close}
-        />
-      )}
-
-      <Header chainId={chainId} onBack={onBack} hideAccount={hideAccount} />
-
+      <Header onBack={onBack} hideAccount={hideAccount} />
       <VStack
         w="full"
         h="full"
-        p={4}
-        pt={12}
         overflowY="auto"
         css={{
           "::-webkit-scrollbar": {
@@ -62,34 +37,15 @@ export function Container({
       >
         {children}
       </VStack>
-
-      <HStack
-        w="full"
-        borderTopWidth={1}
-        borderColor="solid.tertiary"
-        color="text.secondary"
-        alignItems="center"
-        justify="center"
-        h={FOOTER_HEIGHT / 4}
-        bottom={0}
-        position={["fixed", "fixed", "absolute"]}
-        gap={1}
-      >
-        <Text fontSize="xs" color="currentColor">
-          Controller by
-        </Text>
-
-        <CartridgeLogo fontSize={100} color="text.secondary" />
-      </HStack>
     </Wrapper>
   );
 }
 
 export const FOOTER_HEIGHT = 40;
 
-function Wrapper({ children }: React.PropsWithChildren) {
+function Wrapper({ variant = "default", children, ...rest }: React.PropsWithChildren & { variant?: LayoutVariant }) {
   return (
-    <>
+    <LayoutContext.Provider value={{ variant }}>
       {/** Show as full page  */}
       <Show below="md">
         <ChakraContainer
@@ -100,6 +56,7 @@ function Wrapper({ children }: React.PropsWithChildren) {
           animate={{ opacity: 1 }}
           initial={{ opacity: 0 }}
           centerContent
+          {...rest}
         >
           {children}
         </ChakraContainer>
@@ -124,11 +81,24 @@ function Wrapper({ children }: React.PropsWithChildren) {
             borderRadius="md"
             overflow="hidden"
             position="relative"
+            {...rest}
           >
             {children}
           </ChakraContainer>
         </Flex>
       </Show>
-    </>
+    </LayoutContext.Provider>
   );
+}
+
+const LayoutContext = createContext<LayoutContextValue>({ variant: "default" });
+
+type LayoutContextValue = {
+  variant: LayoutVariant
+}
+
+type LayoutVariant = "default" | "connect"
+
+export function useLayoutVariant() {
+  return useContext(LayoutContext).variant
 }
