+++ packages/keychain/src/components/layout/Footer/SessionDetails.tsx
