--- packages/keychain/src/pages/accountQuests.graphql
@@ -1,74 +1,0 @@
-query AccountQuests($accountId: ID!, $gameId: ID) {
-  quests(where: { hasGameWith: { id: $gameId } }) {
-    edges {
-      node {
-        id
-        title
-        description
-        game {
-          id
-        }
-        points
-        metadata {
-          callToAction {
-            text
-            url
-            redirect
-          }
-        }
-        rewards {
-          edges {
-            node {
-              id
-            }
-          }
-        }
-        parent {
-          id
-          title
-        }
-        discordGuild {
-          id
-        }
-        twitterQuests {
-          id
-        }
-        questEvents {
-          id
-          description
-        }
-      }
-    }
-  }
-  account(id: $accountId) {
-    questProgression {
-      edges {
-        node {
-          questID
-          completed
-          claimed
-          completedAt
-          claimTransaction {
-            transactionHash
-          }
-          completion {
-            questEvent
-            completed
-          }
-        }
-      }
-    }
-  }
-}
-
-mutation ClaimQuestRewards($accountId: ID!, $questId: ID!) {
-  claimQuestRewards(accountId: $accountId, questId: $questId)
-}
-
-mutation CheckDiscordQuests($accountId: ID!) {
-  checkDiscordQuests(accountId: $accountId)
-}
-
-mutation CheckTwitterQuests($accountId: ID!) {
-  checkTwitterQuests(accountId: $accountId)
-}
