--- packages/keychain/src/pages/quests.tsx
@@ -1,1 +1,0 @@
-export { Quests as default } from "components";
