--- packages/keychain/src/pages/starterpack.graphql
@@ -1,73 +1,0 @@
-query StarterPack($id: ID!) {
-  game(id: $id) {
-    id
-    name
-    description
-    socials {
-      discord
-      twitter
-      website
-    }
-    icon {
-      uri
-    }
-    profilePicture {
-      uri
-      alt
-    }
-    banner {
-      uri
-      alt
-    }
-    starterPack {
-      id
-      name
-      description
-      chainID
-      issuance
-      maxIssuance
-      starterPackFungibles {
-        contract {
-          id
-          name
-          description
-          priority
-        }
-        amount
-      }
-      starterPackTokens {
-        token {
-          tokenID
-          contract {
-            priority
-          }
-          metadata {
-            name
-            description
-          }
-          thumbnail {
-            uri
-          }
-        }
-        amount
-      }
-      prerequisitesQuests {
-        id
-        title
-        parent {
-          id
-        }
-        metadata {
-          callToAction {
-            text
-            url
-          }
-        }
-      }
-    }
-  }
-}
-
-mutation ClaimStarterpack($id: ID!, $account: ID!) {
-  claimStarterpack(starterpackId: $id, account: $account)
-}
